service error handling and some validations were added to the api

+++ src/main/java/com/ertugrul/springbootmongo/service/CategoryServiceImpl.java
@@ -3,6 +3,7 @@ package com.ertugrul.springbootmongo.service;
 import com.ertugrul.springbootmongo.converter.CategoryConverter;
 import com.ertugrul.springbootmongo.dto.CategoryDto;
 import com.ertugrul.springbootmongo.entity.Category;
+import com.ertugrul.springbootmongo.exception.CategoryNotFoundException;
 import com.ertugrul.springbootmongo.service.entityservice.CategoryEntityService;
 import org.springframework.stereotype.Service;
 
@@ -26,6 +27,8 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto findById(String id) {
         Category category = categoryEntityService.findById(id);
+        if (category == null)
+            throw new CategoryNotFoundException("Category not found. id: " + id);
         return CategoryConverter.INSTANCE.convertCategoryToCategoryDto(category);
     }
 
@@ -37,7 +40,10 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void delete(String id) {
+    public void deleteById(String id) {
+        Category category = categoryEntityService.findById(id);
+        if (category == null)
+            throw new CategoryNotFoundException("Category not found. id: " + id);
         categoryEntityService.deleteById(id);
     }
 }
